docs: update docs and examples
