+++ doc-site/pages/components/form-field.mdx
@@ -10,7 +10,7 @@ import playgroundConfig from '../../examples/form-field/playground-config';
 import BasicUsage from '../../examples/form-field/basic-usage';
 import States from '../../examples/form-field/states';
 import RequiredIndicator from '../../examples/form-field/required-indicator';
-import SmallSize from '../../examples/form-field/small-size';
+import Sizes from '../../examples/form-field/sizes';
 import Inverted from '../../examples/form-field/inverted';
 
 export default Layout;
@@ -35,8 +35,8 @@ Use this utility component in conjunction with [Input](/components/input), [Text
   <RequiredIndicator />
 </Example>
 
-<Example title="Small Variant" path="form-field/small-size">
-  <SmallSize />
+<Example title="Small Variant" path="form-field/sizes">
+  <Sizes />
 </Example>
 
 <Example title="Inverted" path="form-field/inverted" darkPreview>
